Zone detection is pretty good now!
Replaced dilate with blur + threshold, so that random specks don't
create highlighted regions.
Also increased # colors in the KMeans and num iterations...
It's slower now, but more accurate.
And I think getting the KMeans accurate is _really_
important, especially for maps like the Berkeley map where there are
lots of colors that are really close to each other